split dco3 api wrapper
