+++ Cargo.toml
@@ -6,30 +6,28 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+# DRACOON API
+dco3 = { git = "https://github.com/unbekanntes-pferd/dco3.git" }
+
+# CLI helpers
 clap = { version = "4.1.8", features = ["derive"] }
-reqwest = {version = "0.11.14", features = ["json", "stream"]}
-chrono = "0.4.1"
-thiserror = "1.0.2"
-base64 = "0.21.0"
-dotenv = "0.15.0"
-serde = { version = "1.0.159", features = ["derive"] }
-tokio = { version = "1.27.0", features = ["full"] }
-async-trait = "0.1.68"
-tracing = "0.1.37"
-tracing-subscriber = {version = "0.3.17", features = ["env-filter"]}
-serde_json = "1.0.95"
-futures-util = "0.3.28"
-url = "2.3.1"
-serde-xml-rs = "0.6.0"
 indicatif = "0.17.3"
 console = "0.15.5"
 dialoguer = "0.10.4"
-dco3_crypto = "0.5.0"
-tokio-util = { version = "0.7.7", features = ["full"] }
-async-stream = "0.3.5"
-bytes = "1.4.0"
+
+# async runtime and utils
+tokio = { version = "1.27.0", features = ["full"] }
+futures-util = "0.3.28"
+
+# logging and tracing
+tracing = "0.1.37"
+tracing-subscriber = {version = "0.3.17", features = ["env-filter"]}
+
+# utils
+chrono = "0.4.1"
+thiserror = "1.0.2"
+dotenv = "0.15.0"
+
+# credentials
 keyring = "2.0.2"
 
-[dev-dependencies]
-mockito = "1.0.2"
-tokio-test = "0.4.2"
