--- src/api/auth/tests/auth_error.json
@@ -1,4 +1,0 @@
-{
-    "error": "unauthorized",
-    "error_description": "Authentication failed."
-}
