--- src/api/auth/tests/auth_ok.json
@@ -1,8 +1,0 @@
-{
-    "access_token": "access_token",
-    "token_type": "bearer",
-    "refresh_token": "refresh_token",
-    "expires_in_inactive": 28800,
-    "expires_in": 3600,
-    "scope": "all"
-}
